realized that miles should actually be int, so changed int to test for int when converting cash -> miles

+++ src/main/java/RewardValue.java
@@ -19,7 +19,7 @@ public class RewardValue {
     }
 
     private static double convertToMiles(double cashValue){
-        return  cashValue/ MILES_TO_CASH_CONVERSION_RATE;
+        return  (int)(cashValue/ MILES_TO_CASH_CONVERSION_RATE);
     }
 
     private static double convertToCash(int milesValue){
